requirements.txt
pandas>=1.3.0
numpy>=1.20.0
yfinance>=0.1.70
scipy>=1.7.0
pytest>=6.0.0
pyyaml>=5.4.0 